Вынес операции в отдельные файлы,поменял цвета.
