+++ fintech_chat/fintech_chat/Model/DataManager/Operation/Operations/LoadDataOperation.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+class LoadDataOperation: Operation{
+    var url: URL
+    var globalError = false
+    
+    init(url: URL) {
+        self.url = url
+    }
+}
