+++ fintech_chat/fintech_chat/Model/DataManager/Operation/Operations/LoadImageOperation.swift
@@ -1,0 +1,17 @@
+import UIKit
+
+class LoadImageOperation: LoadDataOperation{
+    var imageResult: UIImage? = nil
+    
+    override func main() {
+        guard !isCancelled else {return}
+        
+        do{
+            let data = try Data(contentsOf: url)
+            imageResult = UIImage(data: data)
+        }
+        catch{
+            Logger.app.logMessage("Cant load image from file. \(error.localizedDescription)", logLevel: .Error)
+        }
+    }
+}
