+++ fintech_chat/fintech_chat/Model/DataManager/Operation/Operations/SaveDataOperation.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+class SaveDataOperation: Operation{
+    var url: URL
+    
+    var globalError = false
+    
+    init(url: URL){
+        self.url = url
+    }
+}
+
+
