+++ fintech_chat/fintech_chat/Model/DataManager/Operation/Operations/SaveImageOperation.swift
@@ -1,0 +1,29 @@
+import UIKit
+
+class SaveImageOperation: SaveDataOperation{
+    var imageData: UIImage
+    
+    init(url: URL, imageData: UIImage) {
+        self.imageData = imageData
+        super.init(url: url)
+    }
+    
+    override func main() {
+        guard !isCancelled else {
+            Logger.app.logMessage("Cancel operation", logLevel: .Error)
+            return
+        }
+        do{
+            if let data = imageData.pngData(){
+                try data.write(to: url)
+            }
+            else{
+                globalError = true
+            }
+        }
+        catch{
+            Logger.app.logMessage("Can`t write description to file. \(error.localizedDescription)", logLevel: .Error)
+            globalError = true
+        }
+    }
+}
