+++ fintech_chat/fintech_chat/Style/Controls/AppTextView.swift
@@ -1,0 +1,3 @@
+import UIKit
+
+class AppChatTextView: UITextView{}
